[Ref] Move FullTextSearchResult into the search package in dtos.
