+++ src/main/java/cz/cvut/kbss/termit/dto/search/FullTextSearchResult.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with this program.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package cz.cvut.kbss.termit.dto;
+package cz.cvut.kbss.termit.dto.search;
 
 import cz.cvut.kbss.jopa.model.annotations.*;
 import cz.cvut.kbss.jopa.vocabulary.RDFS;
